requirements-dev.txt
pytest==6.2.5
maturin>=0.13.0<14
pre-commit==2.16.0
polars==0.16.1
testing.postgresql==1.3.0
psycopg==3.1.8
